requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
beautifulsoup4==4.7.1
black==19.3b0
Click==7.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.3.1
python-scriptures==3.0.0
six==1.12.0
soupsieve==1.9.1
toml==0.10.0
typed-ast==1.3.5
wrapt==1.11.1
